% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_achievement.R
\name{adjusted_logistic}
\alias{adjusted_logistic}
\title{A modified logistic function to use for achievement assessment with thresholds}
\usage{
adjusted_logistic(x, threshold, k, mid = 0, thresh_is)
}
\arguments{
\item{x}{the input value}

\item{threshold}{the threshold to compare against}

\item{k}{the k of the logistic}

\item{mid}{the midpoint of logistic, default 0 \emph{be very careful if change}}

\item{thresh_is}{is the threshold a 'min' or a 'max', e.g. maximum interevents are a 'max', frequency requirements are a 'min'}
}
\value{
a value
}
\description{
A modified logistic function to use for achievement assessment with thresholds
}
