% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_colors.R
\name{as_colors}
\alias{as_colors}
\title{Make a vector of colors have class 'colors'}
\usage{
as_colors(colvec)
}
\arguments{
\item{colvec}{character vector of colors}
}
\value{
object of class colors
}
\description{
This \emph{must} exist somewhere, because other packages make things with class 'colors', but when I c() to them it breaks.
}
\keyword{internal}
