% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_ewr_output.R
\name{assess_ewr_achievement}
\alias{assess_ewr_achievement}
\title{EWR logic test on incoming Annual EWRs}
\usage{
assess_ewr_achievement(
  annualdf,
  year_roll = ifelse(nrow(annualdf) >= 10, 10, 1)
)
}
\arguments{
\item{annualdf}{incoming tibble of EWRs after read-in}

\item{year_roll}{specific number of years to check assessment for}
}
\value{
tibble with columns \code{scenario}, \code{year}, \code{date}, \code{gauge}, \code{planning_unit_name}, \code{state}, \code{SWSDLName}, \code{ewr_code}, \code{ewr_code_main}, \code{event_years}, \code{frequency_achieved}, \code{interevent_achieved}, \code{ewr_achieved}
}
\description{
This is the pass/fail test of whether the criteria (frequency and timing) of EWRs are met or not
Using the Minimum long term average (LTA) target frequencies (Termed Target frequency in summarydf) as suggested in the LTWPs
Includes inverse result for cease to flows (CF) in both the assessment and the event_years themselves
}
