% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_achievement.R
\name{assess_ewr_flex}
\alias{assess_ewr_flex}
\title{Example of a fairly complex function to prepare data pre-aggregation (as a \code{prepfun} argument to \code{\link[=read_and_agg]{read_and_agg()}})}
\usage{
assess_ewr_flex(annualdf, year_roll, k_freq, k_inter, combo)
}
\arguments{
\item{combo}{}
}
\description{
\emph{This should be considered a template, and does not have general functionality}
}
