% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{basin}
\alias{basin}
\title{Murray-Darling Basin Polygon}
\format{
\subsection{\code{basin} An \code{sf} object with one polygon:}{

\describe{
\item{DDIV_NAME}{Polygon name}
\item{AREA_HA, SHAPE_AREA, SHAPE_LEN}{basic descriptions}
}
}
}
\source{
\url{https://data.gov.au/data/dataset/4ede9aed-5620-47db-a72b-0b3aa0a3ced0/resource/8a6d889d-723b-492d-8c12-b8b0d1ba4b5a/download/sworkingadhocjobsj4430dataoutputsmdb_boundarymdb_boundary.zip}
}
\usage{
basin
}
\description{
Polygon defining the Murray-Darling Basin boundary. Sourced from \href{https://www.mdba.gov.au/publications-and-data/maps-and-spatial-data/spatial-data}{MDBA data website},
with direct link given below.
}
\keyword{datasets}
