% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_ewr_output.R
\name{bind_max}
\alias{bind_max}
\title{Add max scenario}
\usage{
bind_max(outdf)
}
\arguments{
\item{outdf}{#EWR summary output with pass fail results}
}
\value{
tibble with a MAX scenario added
}
\description{
Add max scenario
}
\keyword{internal}
