% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{causal_ewr_simple}
\alias{causal_ewr_simple}
\title{Simplified causal relationships for the EWRs at the basin scale}
\format{
\subsection{\code{causal_ewr} A list of dataframes:}{

\describe{
\item{ewr2target}{Mapping from EWRs at each gauge to \code{Target} group}
\item{target2yrtarget}{Mapping defined without spatial reference from \code{Target}, to \code{Target_Category}, \code{Objective}, \code{target_5_year_2024}, \code{target_10_year_2029}, and \code{target_20_year_2039}, based \emph{only} on NSW at present}
}
}
}
\source{
various Long-Term Watering Plans and State-based tables.
}
\usage{
causal_ewr_simple
}
\description{
List of dataframes specifying the causal relationships from environmental
water requirements  to Target groups. Skips over the environmental objectives
(e.g. the individual species or life cycle components) because the links from
those to the Targets are not currently available everywhere. If working in
areas where they \emph{are} available, use \code{causal_ewr}. Extracted and cleaned in
\code{data_creation/causal_data_cleaning.qmd}, with definitions found and
extracted from the EWR tool. The links in target2yrtarget are from NSW and
so, while general, have not been confirmed to apply in other states.
}
\keyword{datasets}
