% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cewo_valleys}
\alias{cewo_valleys}
\title{Catchments within the Murray-Darling Basin}
\format{
\subsection{\code{cewo_valleys} An {sf} object:}{

\describe{
\item{ValleyName, ValleyID, ValleyCode}{Name, ID, and code for each catchment}
}
}
}
\source{
\url{https://data.gov.au/data/dataset/75910bc5-6c3e-40e8-9c8a-1e895274badb/resource/a7053ee7-8e20-4f2c-b594-cb88c6ed9406/download/cewo_mdb_valleys.shp.zip}
}
\usage{
cewo_valleys
}
\description{
Polygons defining catchment boundaries, as defined by the CEWO. Sourced from
\href{https://www.mdba.gov.au/publications-and-data/maps-and-spatial-data/spatial-data}{MDBA data website},
with direct link given below. The 'Northern Unregulated' polygon has been
removed, as it is the combination of several other polygons.
}
\keyword{datasets}
