% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory_setup.R
\name{check_missing_runs}
\alias{check_missing_runs}
\title{Similar to find_missing_runs, but excised from prep_run_save_ewrs to take the same arguments}
\usage{
check_missing_runs(
  hydro_dir,
  output_parent_dir,
  output_subdir = "",
  scenarios = NULL,
  file_search = NULL,
  model_format = "Standard time-series",
  outputType = "none",
  scenarios_from = "directory"
)
}
\arguments{
\item{hydro_dir}{Directory containing hydrographs. Can be an outer directory,
e.g. \code{hydrographs} that splits into scenario subdirs, or can be a single
scenario subdir.}

\item{output_parent_dir}{parent directory for the outputs. Can be anything,
but two typical cases:
\itemize{
\item The directory containing \code{hydro_dir}, which puts the \code{module_outputs} at the same level as the hydrographs
\item If running in batches for single scenarios, may be \code{hydro_dir}, which just puts the \code{module_outputs} in \code{hydro_dir}
}}

\item{output_subdir}{a sub-directory for the outputs, if for example we want
\code{module_output/EWR/V1} and \code{module_output/EWR/V2}}

\item{scenarios}{\code{NULL} (default) or named list.
\itemize{
\item \code{NULL}- finds scenario names by parsing directory names in \code{hydro_dir}. If no internal directories, just stays in \code{hydro_dir}. This captures the two typical situations discussed for \code{output_parent_dir}. If there are other directories in \code{hydro_dir} that do not contain hydrological scenarios, should use a character vector.
\item named list of paths to files. names become scenario names, paths should be relative to \code{hydro_dir}. This allows unusual directory structures.
}}

\item{file_search}{character, regex for additional limitations on filenames.
Useful to run a subset of scenarios or if several files have the extension
defined by \code{model_format}, but only some are hydrographs.}

\item{model_format}{see EWR tool. One of:
\itemize{
\item 'Standard time-series': (default, among other things accepts a csv with a Date column followed by gauge columns, with _flow or _level appended to the gauge number)
\item 'IQQM - netcdf': in development, finds all netcdf files in \code{hydro_dir}. Should also work when \code{hydro_dir} is a .zip with netcdfs inside
\item 'ten thousand year': old default (IQQM - NSW 10,000 years), works nearly the same as standard time-series
\item 'All Bigmod': previously 'Bigmod - MDBA'
\item 'Source - NSW (res.csv)'
}}

\item{outputType}{list of strings or character vector defining what to save
to disk. One or more of:
\itemize{
\item 'none' (default), do not save outputs- ignored if in a list with others
\item 'summary',
\item 'yearly',
\item 'all_events',
\item 'all_successful_events',
\item 'all_interEvents', # Does not work with EWR 2.0 tool
\item 'all_successful_interEvents'
}}

\item{scenarios_from}{character, default 'directory' gets scenario names from
directory names. If anything else, gets them from filenames (safest).
Expect additional options in future, e.g from metadata.}
}
\value{
character vector of missing filepaths
}
\description{
Similar to find_missing_runs, but excised from prep_run_save_ewrs to take the same arguments
}
