% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_ewr_causal.R
\name{clean_ewr_237}
\alias{clean_ewr_237}
\title{Clean causal network for EWR tool 2.3.7 and thereabouts}
\usage{
clean_ewr_237(ewrnet)
}
\arguments{
\item{ewrnet}{}
}
\description{
Clean causal network for EWR tool 2.3.7 and thereabouts
}
