% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_ewr_causal.R
\name{clean_ewr_causal}
\alias{clean_ewr_causal}
\title{Cleans causal networks that come in from py-ewr}
\usage{
clean_ewr_causal(ewrnet, verbose = FALSE)
}
\arguments{
\item{ewrnet}{causal network returned by pdi$get_causal_ewr()}

\item{verbose}{default FALSE, if TRUE, print out a bunch of unique values for troubleshooting}
}
\value{
a list of dataframes
}
\description{
Cleans causal networks that come in from py-ewr
}
\keyword{internal}
