% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_run_save_ewrs.R
\name{clean_ewr_in_R}
\alias{clean_ewr_in_R}
\title{Clean up the EWR outputs so returns look like they do when read from csv. Mostly dateparsing from py objects}
\usage{
clean_ewr_in_R(ewr_out)
}
\arguments{
\item{ewr_out}{}
}
\value{
tibble, cleaned EWR outputs
}
\description{
Clean up the EWR outputs so returns look like they do when read from csv. Mostly dateparsing from py objects
}
