% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ewr_output.R
\name{clean_ewr_requirements}
\alias{clean_ewr_requirements}
\title{Cleans the EWR table to use for assessing target requirements by the outputs}
\usage{
clean_ewr_requirements()
}
\value{
clean tibble suitable for joining to outputs
}
\description{
Cleans the EWR table to use for assessing target requirements by the outputs
}
