% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_long_term.R
\name{clean_long_term}
\alias{clean_long_term}
\title{Process objective-long term target links}
\usage{
clean_long_term(yrpath, saveout = FALSE, outdir, savename)
}
\arguments{
\item{yrpath}{path to the file with matched objectives and targets}

\item{saveout}{logical. saves a file of the output as \code{outdir/savename_YearMonthDayHourMinute.csv}}

\item{outdir}{path to the directory to save into. not needed if \code{saveout} is \code{FALSE}}

\item{savename}{one of FALSE, 'r', or 'csv', controls saving and type.
\itemize{
\item \code{FALSE}: (the default): Does not save anything
\item \code{'r'}: saves an rdata file, used primarily for building the data directory in a package
\item \code{'csv'}: saves a csv with the name \code{outdir/savename_YearMonthDayHourMinute.csv}
}}
}
\value{
a \code{tibble} with matching env_objs, Targets, Objectives, and the 5,10, and 20 year targets. No spatial information (gauge, planning unit)
}
\description{
DEPRECATED- use \code{\link[=get_causal_ewr]{get_causal_ewr()}}. Bespoke function to clean up the current input data
}
