% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_obj_target.R
\name{clean_obj_target}
\alias{clean_obj_target}
\title{Pair environmental objectives to various targets and objectives}
\usage{
clean_obj_target(
  ewrobjs,
  targetpath,
  qcfiles,
  saveout = FALSE,
  outdir,
  savename
)
}
\arguments{
\item{ewrobjs}{\code{tibble} or \code{data.frame} of the EWR indicators paired to the environmental objectives. Should be the "long" output from \code{clean_ewr_obj}}

\item{targetpath}{path to the table linking objectives to targets}

\item{qcfiles}{list of paths to QC files from Renee}

\item{saveout}{one of FALSE, 'r', or 'csv', controls saving and type.
\itemize{
\item \code{FALSE}: (the default): Does not save anything
\item \code{'r'}: saves an rdata file, used primarily for building the data directory in a package
\item \code{'csv'}: saves a csv with the name \code{outdir/savename_YearMonthDayHourMinute.csv}
}}

\item{outdir}{path to the directory to save into. not needed if \code{saveout} is \code{FALSE}}

\item{savename}{character for the filename to save. the date and time gets appended to avoid overwriting}
}
\value{
a \code{tibble} linking the \code{env_obj} to several different outcomes- \code{Specific_goal} (roughly species, but also things like refugia), \code{Objective} (bigger-picture, e.g. no loss of native fish species), and \code{Target} (Native Fish, Ecosystem Function, etc). At the Planning Unit scale.
}
\description{
DEPRECATED- use \code{\link[=get_causal_ewr]{get_causal_ewr()}}.Bespoke function to clean up the current input data.
}
\keyword{internal}
