% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_hydrobot_params.R
\name{clean_sequences}
\alias{clean_sequences}
\title{Clean up the sequences if they are provided in r scripts or otherwise}
\usage{
clean_sequences(arglist)
}
\arguments{
\item{arglist}{}
}
\value{
arglist
}
\description{
Clean up the sequences if they are provided in r scripts or otherwise
}
\keyword{internal}
