% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ewr_output.R
\name{clean_yearly}
\alias{clean_yearly}
\title{Additional cleanup specific to yearly EWR outputs.}
\usage{
clean_yearly(annualdf)
}
\arguments{
\item{annualdf}{the result of \code{\link[=get_any_ewr_output]{get_any_ewr_output()}} with \code{type = 'yearly'}}
}
\description{
Additional cleanup specific to yearly EWR outputs.
}
