% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_ewr_output.R
\name{cleanewrs}
\alias{cleanewrs}
\title{Clean up and standardise names and column types}
\usage{
cleanewrs(ewrdf)
}
\arguments{
\item{ewrdf}{ewr dataframe of any type}
}
\value{
a tibble of EWR outputs with cleaned up names, separated ewr_code_main, and gauges as characters
}
\description{
The new format of the EWR output makes this much lighter than previously.
}
