% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ewr_output.R
\name{cleanSummary}
\alias{cleanSummary}
\title{Clean the incoming summary EWRs}
\usage{
cleanSummary(summarydf)
}
\arguments{
\item{summarydf}{incoming tibble of EWRs after read-in}
}
\value{
tibble reformatted and cleaned for ongoing analysis
}
\description{
There is some data organisation that happens for the toolkit in the scenario
controller, but this does a bit more cleaning of data format for further
analysis
}
