% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_achievement.R
\name{control_achievement}
\alias{control_achievement}
\title{An example of a \code{prepfun} for \code{\link[=read_and_agg]{read_and_agg()}}}
\usage{
control_achievement(dat, year_roll = 10, k_freq, k_inter, combo)
}
\arguments{
\item{combo}{}
}
\description{
\emph{This should be considered a template, and does not have general functionality}
}
