% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauge_pu.R
\name{extract_gauge_pu}
\alias{extract_gauge_pu}
\title{Extract the gauge to planning unit match from a dataframe.}
\usage{
extract_gauge_pu(rellist)
}
\arguments{
\item{rellist}{a dataframe or tibble or a list of dataframes/tibbles. At
least one of the dataframes needs to have a \code{"gauge"} column and a
\code{"PlanningUnitID"} column}
}
\value{
a tibble with the PlanningUnitID for each gauge in \code{rellist}
}
\description{
This is not spatial, it gets the relationship from a passed in df. This way
ensures internal consistency when dealing with various levels of causal
relationships
}
