% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{find_color_type}
\alias{find_color_type}
\title{Helper to find the type of color specification, so we can call the right \verb{ggplot::scale_color_*}}
\usage{
find_color_type(pal_list)
}
\arguments{
\item{pal_list}{as in \code{\link[=plot_outcomes]{plot_outcomes()}}}
}
\value{
character, one of 'colorobj', 'paletteer_c', 'paletteer_d', 'grouped', 'fixed' that help other functions know what to do wiht color/fill
}
\description{
Four options:
\enumerate{
\item pal_list is a named \code{colors} object
return an indication to use scale_*manual(values = pal_list)
\item pal_list is a paletteer palette name
return an indication to use scale_paletteer_*(pal_list)
\item pal_list is a list, and we need to deal with grouped colors
find the colors to use, return those in the data along with an indication to use scale_*identity
\item pal_list is a single color value
Does not automate other things like making the colors
object from data and pal_list; not necessary and is fraught.
}
}
\keyword{internal}
