% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory_setup.R
\name{find_expected_files}
\alias{find_expected_files}
\title{Figure out the expected outputs}
\usage{
find_expected_files(hydro_paths, output_path, outputType, scenarios_from)
}
\arguments{
\item{hydro_paths}{list of paths to scenario files with names scenario names}

\item{output_path}{path for outputs}

\item{outputType}{types of outputs, for inferring csvs}

\item{scenarios_from}{as in \code{\link[=prep_run_save_ewrs]{prep_run_save_ewrs()}}}
}
\value{
character vector of expected filepaths
}
\description{
Figure out the expected outputs
}
