% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{find_limits}
\alias{find_limits}
\title{Finds limits for y-axis and colors}
\usage{
find_limits(limcol, lims, trans, base_list)
}
\arguments{
\item{lims}{desired limits. see \code{setLimits} in \code{\link[=plot_outcomes]{plot_outcomes()}}}

\item{trans}{transform if any}

\item{base_list}{as in \code{\link[=plot_outcomes]{plot_outcomes()}}}

\item{ycol}{vector of the y-values, typically \code{sf::st_drop_geometry(prepped$data[prepped$outcome_col])}}
}
\value{
length-2 limits (or NULL)
}
\description{
Very similar to \code{\link[=findlimits]{findlimits()}}, which has been deprecated
}
