% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory_setup.R
\name{find_scenario_paths}
\alias{find_scenario_paths}
\title{get the paths to scenarios}
\usage{
find_scenario_paths(hydro_dir, type = "csv", file_search = NULL)
}
\arguments{
\item{hydro_dir}{directory with hydrographs}

\item{type}{filetype, default 'csv', likely will also need 'ncdf'}

\item{file_search}{character, regex for additional limitations on filenames. Useful if several files have the extension defined by \code{type}, but only some are hydrographs.}
}
\value{
a list of filepaths to the hydrographs
}
\description{
get the paths to scenarios
}
