% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{findlimits}
\alias{findlimits}
\title{Finds limits for the color scale, accounting for prepped data and baselining.}
\usage{
findlimits(x, lims, base_list)
}
\arguments{
\item{x}{the default data range passed in by the palette}

\item{lims}{desired limits}

\item{base_list}{as in \code{\link[=plot_outcomes]{plot_outcomes()}}}
}
\description{
allows (by crude inference) centering diverging palettes with baseline comparisons
}
