% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{findrescale}
\alias{findrescale}
\title{Set the midpoint of the color/fill. DANGEROUS- DO NOT USE WITHOUT A LOT MORE TESTING}
\usage{
findrescale(
  x,
  lims,
  base_list,
  from = range(x, na.rm = TRUE),
  trans = "identity"
)
}
\arguments{
\item{x}{data}

\item{lims}{the seLimits, see \code{\link[=plot_outcomes]{plot_outcomes()}}}

\item{base_list}{see \code{\link[=plot_outcomes]{plot_outcomes()}}}

\item{from}{passed argument, have to handle}

\item{to}{passed argument, have to handle}
}
\value{
vector
}
\description{
Set the midpoint of the color/fill. DANGEROUS- DO NOT USE WITHOUT A LOT MORE TESTING
}
