% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionlister.R
\name{functionlister}
\alias{functionlister}
\title{Function cleanup to allow passing character, bare names, and list-formatted functions}
\usage{
functionlister(funs, forcenames = NULL)
}
\arguments{
\item{funs}{a function to evaluate inside something else, could be character, bare name, or list}
}
\description{
Function cleanup to allow passing character, bare names, and list-formatted functions
}
