% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauge_pu.R
\name{gauge_ltwp_ewr}
\alias{gauge_ltwp_ewr}
\title{Extract the ewrs for each gauge, along with the PlanningUnitID and
LTWPShortName}
\usage{
gauge_ltwp_ewr()
}
\value{
dataframe with \code{PlanningUnitID}, \code{LTWPShortName}, \code{gauge}, and
\code{ewr_code} columns as they exist in the good EWR table from py-ewr
}
\description{
Pulls straight from py-ewr, so very useful for various cross-referencing and
keeping up to date with the state of that package.
}
\examples{
gauge_ltwp_ewr()
}
