% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauge_pu.R
\name{gauge_pu}
\alias{gauge_pu}
\title{Get matches between gauges and planning units Based entirely on the EWR
dataframe, NOT actual spatial referencing. That will happen elsewhere. This
way ensures internal consistency when dealing with various levels of causal
relationships}
\usage{
gauge_pu(rellist)
}
\arguments{
\item{rellist}{a dataframe or tibble or a list of dataframes/tibbles. At least one of the dataframes needs to have a \code{"gauge"} column and a \code{"PlanningUnitID"} column}
}
\value{
a tibble with the PlanningUnitID for each gauge in \code{rellist}
}
\description{
Get matches between gauges and planning units Based entirely on the EWR
dataframe, NOT actual spatial referencing. That will happen elsewhere. This
way ensures internal consistency when dealing with various levels of causal
relationships
}
