% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ewr_output.R
\name{get_any_ewr_output}
\alias{get_any_ewr_output}
\title{Get the outputs from the EWR and clean them up}
\usage{
get_any_ewr_output(dir, type, gaugefilter = NULL, scenariofilter = NULL)
}
\arguments{
\item{dir}{path to directory with the EWR output for all gauges and scenarios}

\item{type}{character, one of
\itemize{
\item 'summary',
\item 'yearly',
\item 'all_events',
\item 'all_successful_events',
\item 'all_interEvents', # Does not work with current EWR tool
\item 'all_successful_interEvents'
}}

\item{gaugefilter}{character vector of gauge numbers to include. Default \code{NULL} includes all}

\item{scenariofilter}{character vector of scenario names to include. Default \code{NULL} includes all}
}
\value{
a tibble of EWR results if \code{type == "annual"} or \code{type == "summary"}. A named list of those tibbles if \code{type == "both"}
}
\description{
\emph{NOTE} this expects output from the EWRs to be processed through HydroBOT
functions for cleaning and saving immediately after creation, which may
change now that \code{py-ewr} has changed some things.
}
