% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ewr_info.R
\name{get_causal_ewr}
\alias{get_causal_ewr}
\title{Extract causal network from EWR tool}
\usage{
get_causal_ewr(struct = "list")
}
\value{
list of EWR causal networks
}
\description{
Extract causal network from EWR tool
}
