% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plot_table.R
\name{get_data_agg}
\alias{get_data_agg}
\title{The data aggregation extractor for make_plot_table}
\usage{
get_data_agg(werp_plot)
}
\arguments{
\item{werp_plot}{A ggplot object with WERP-style data underneath. Needs to have the aggregation sequence in columns, not the column name.}
}
\description{
The data aggregation extractor for make_plot_table
}
\keyword{internal}
