% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ewr_info.R
\name{get_ewr_gauges}
\alias{get_ewr_gauges}
\title{get the gauges from the EWR tool}
\usage{
get_ewr_gauges()
}
\value{
sf with unique gauges with good EWRs
}
\description{
get the gauges from the EWR tool
}
