% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ewr_output.R
\name{get_ewr_output}
\alias{get_ewr_output}
\title{Make EWR results with achievement for ongoing use}
\usage{
get_ewr_output(dir, year_roll = 1)
}
\arguments{
\item{dir}{path to directory with the EWR output for all gauges and scenarios}

\item{year_roll}{}
}
\value{
a tibble with ewr_achieved
}
\description{
Make EWR results with achievement for ongoing use
}
