% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ewr_info.R
\name{get_ewr_table}
\alias{get_ewr_table}
\title{Get the table of EWRs from the EWR tool}
\usage{
get_ewr_table(type = "good")
}
\arguments{
\item{type}{'good', default, or 'bad'}
}
\value{
dataframe of the ewr table
}
\description{
Get the table of EWRs from the EWR tool
}
