% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ewr_info.R
\name{get_ewr_version}
\alias{get_ewr_version}
\title{Extract the package version. Does not get complications like git branches.}
\usage{
get_ewr_version()
}
\value{
character version number of the py-ewr package
}
\description{
Extract the package version. Does not get complications like git branches.
}
