% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ewr_info.R
\name{get_iqqm_gauges}
\alias{get_iqqm_gauges}
\title{Get the mapping of gauges to IQQM nodes for netcdf}
\usage{
get_iqqm_gauges()
}
\value{
dataframe of the ewr table
}
\description{
Get the mapping of gauges to IQQM nodes for netcdf
}
