% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped_colours.R
\name{grouped_colors}
\alias{grouped_colors}
\title{Sets up colors within groups according to separate palettes}
\usage{
grouped_colors(
  df,
  pal_list,
  pal_direction = rep(1, length(pal_list)),
  colorgroups = NULL,
  colorset = NULL,
  setLimits = NULL
)
}
\arguments{
\item{df}{input dataframe}

\item{pal_list}{list of palettes, or color name or color object that maps to \code{colorgroups}, \code{colorset}, or \code{scenario}. Typically a list of names of {palletteer} palettes of the same length as unique values in \code{colorgroups}}

\item{colorgroups}{character column name of column defining the groups that define different palettes to use}

\item{colorset}{character column name of the values to assign colors from those palettes}

\item{setLimits}{NULL (default) or length-2 numeric vector to force limits of the color scale.}
}
\value{
a dataframe with new column \code{color} with hex colors
}
\description{
Sets up colors within groups according to separate palettes
}
