% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation_helpers.R
\name{is_notpoint}
\alias{is_notpoint}
\title{Test whether an object is an sf other than a point}
\usage{
is_notpoint(x)
}
\arguments{
\item{x}{dataframe or other object}
}
\value{
logical, TRUE if sf and not all data are POINTs
}
\description{
This is not just !is_point(), because it still needs to be an sf.
}
