% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation_helpers.R
\name{is_point}
\alias{is_point}
\title{Test whether an object is a spatial point sf}
\usage{
is_point(x)
}
\arguments{
\item{x}{dataframe or other object}
}
\value{
logical, TRUE if is sf and all data are POINTs
}
\description{
The data has to be both an sf and then only contain points
}
