% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation_helpers.R
\name{is_sf}
\alias{is_sf}
\title{Test whether an object is an sf}
\usage{
is_sf(x)
}
\arguments{
\item{x}{anything, but typically a dataframe}
}
\value{
Logical
}
\description{
Test whether an object is an sf
}
\examples{
is_sf(basin)
is_sf(iris)
}
