% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation_helpers.R
\name{is_theme}
\alias{is_theme}
\title{Test whether a list-item specifies a theme dimension}
\usage{
is_theme(x, causalnames)
}
\arguments{
\item{x}{list-item}

\item{causalnames}{names of the columns in a causal network}
}
\value{
logical
}
\description{
Intended to be purrr-ed over from aggsequence
}
\keyword{internal}
