% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation_helpers.R
\name{is_time}
\alias{is_time}
\title{Test whether a list-item is time, including using \code{\link[base:cut.POSIXt]{base::cut.Date()}} specifications}
\usage{
is_time(x)
}
\arguments{
\item{x}{a list-item}
}
\value{
Logical
}
\description{
Intended to be purrr-ed over from aggsequence
}
