% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_to_geo.R
\name{join_to_geo}
\alias{join_to_geo}
\title{Pairs gauge numbers with geographic locations to make an \code{sf}}
\usage{
join_to_geo(data_df, spatial_locs, whichcrs = 4283)
}
\arguments{
\item{data_df}{A tibble or dataframe with a \code{gauge} column with gauge numbers as characters}

\item{spatial_locs}{sf or path to the csv with gauge numbers paired with lat-long.}

\item{whichcrs}{desired crs of the output. input crs assumed to be WGS84, EPSG 4326}
}
\value{
an \code{sf} dataframe with gauge locations as points in the geometry
}
\description{
Typically used for EWR outputs, but anything with a \code{gauge} column with gauge numbers as characters will work.
}
