% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_agg_funs.R
\name{LimitingFactor}
\alias{LimitingFactor}
\title{Limiting aggregation}
\usage{
LimitingFactor(x, na.rm = TRUE)
}
\arguments{
\item{x}{values to aggregate}

\item{na.rm}{default \code{na.rm = TRUE} finds the aggregation ignoring NA, opposite usual na.rm default \code{na.rm = FALSE}, which returns NA if there are any NA.}
}
\value{
summarised data
}
\description{
Doing poorly at one time/place negates benefits elsewhere. Implemented with min(). For binary, this means failing anywhere means failing everywhere.
}
