% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_causal_plot.R
\name{make_causal_plot}
\alias{make_causal_plot}
\title{Builds a causal network plot}
\usage{
make_causal_plot(
  nodes,
  edges,
  focalnodes = unique(nodes$Name),
  drop_unused_nodes = TRUE,
  edge_pal = list(fromtype = "nationalparkcolors::GeneralGrant"),
  edge_colorgroups = NULL,
  edge_colorset = "fromtype",
  edge_pal_direction = rep(1, length(edge_pal)),
  node_pal = list(fromtype = "nationalparkcolors::GeneralGrant"),
  node_colorgroups = NULL,
  node_colorset = "NodeType",
  node_pal_direction = rep(1, length(node_pal)),
  setLimits = NULL,
  wrap_names = TRUE,
  render = TRUE,
  returnnetwork = TRUE,
  save = FALSE,
  savedir,
  savename = NULL
)
}
\arguments{
\item{nodes}{a tibble or dataframe defining nodes}

\item{edges}{a tibble or dataframe defining all edge relationships}

\item{focalnodes}{a character vector of nodes to use to find the related
network (default is all values in the \code{Name} column of the \code{nodes} df).
Uses \code{find_related_nodes} to build a network with these nodes along with
all their upstream and downstream nodes.}

\item{drop_unused_nodes}{logical, default \code{TRUE}, drops nodes with no
connections. \code{FALSE} could be useful to keep unconnected nodes if we want
to see if there are any that are not causally related to the others.}

\item{edge_pal}{a named list of color palettes or color values for the
edges.  Currently only supports paletteer palettes, vectors of colors, or
single colors. See \code{\link{causal_colors_general}}}

\item{edge_colorgroups}{NULL (the default) or a column name (character)
specifying a grouping column for different color palettes for the edges.
See \code{\link{causal_colors_general}}}

\item{edge_colorset}{a column name (character) giving the column to define
edge color. See \code{\link{causal_colors_general}}}

\item{edge_pal_direction}{either 1 (default) or -1 (reversed) direction of the palettes}

\item{node_pal}{a named list of color palettes or color values for the
nodes. Currently only supports paletteer palettes, vectors of colors, or
single colors. See \code{\link{causal_colors_general}}}

\item{node_colorgroups}{NULL (the default) or a column name (character)
specifying a grouping column for different color palettes for the nodes.
See \code{\link{causal_colors_general}}}

\item{node_colorset}{a column name (character) giving the column to define
node color. See \code{\link{causal_colors_general}}}

\item{node_pal_direction}{either 1 (default) or -1 (reversed) direction of the palettes}

\item{setLimits}{NULL (default) or length-2 numeric vector to force limits of the color scale.}

\item{wrap_names}{logical, default \code{TRUE}, should long node names be wrapped
to multiple lines}

\item{render}{logical, default \code{TRUE}, should the network be rendered to the
graphics device?}

\item{returnnetwork}{logical, default \code{TRUE}, should the graph object
produced by \code{DiagrammeR} be returned?}

\item{save}{logical, default \code{FALSE}, should the graph be saved as a .png and
.pdf}

\item{savedir}{path to directory for saving outputs. Only needed if \code{save = TRUE}}

\item{savename}{character, filename for the output plots.
.pdf and .png are automatically appended.}
}
\value{
causal network in DiagrammeR format
}
\description{
Takes node and edges dfs, along with arguments to specify attributes
(particularly color), and how to return the output. Internally, this finds
the network relevant to a given set of nodes (by default all of them), using
\code{find_related_nodes}, attributes according to the attribute arguments using
\code{causal_colors_general} and \code{node_plot_atts} and then uses \code{DiagrammeR} to
make the plot.
}
