% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_toolkit_params.R
\name{make_default_args}
\alias{make_default_args}
\title{Helper to allow making the directory structure programatically from the base directory}
\usage{
make_default_args(arglist)
}
\arguments{
\item{arglist}{}
}
\description{
Helper to allow making the directory structure programatically from the base directory
}
