% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_edges.R
\name{make_edges}
\alias{make_edges}
\title{Create edges dataframe for causal network or theme aggregation}
\usage{
make_edges(
  dflist,
  fromtos,
  fromfilter = NULL,
  tofilter = NULL,
  gaugefilter = NULL,
  pufilter = NULL,
  gaugeplanmatch = NULL,
  extrasave = NULL
)
}
\arguments{
\item{dflist}{list of tibbles or dataframes containing the matches between
nodes. Single dataframes may hold multiple relationships. List of
dataframes because some relationships are at different scales}

\item{fromtos}{list of length-2 character vectors \code{c('from', 'to')}, giving
the column names of the node categories to connect (and implies
directionality).}

\item{fromfilter}{character vector of nodes to include (a subset of the nodes
in the \code{from} column. Default \code{NULL} includes all nodes within the column.}

\item{tofilter}{character vector of nodes to include (a subset of the nodes
in the \code{to} column. Default \code{NULL} includes all nodes within the column}

\item{gaugefilter}{character vector of gauge numbers to include. Default
\code{NULL} includes all gauges}

\item{pufilter}{character vector of planning units to include. Default \code{NULL}
includes all planning units}

\item{gaugeplanmatch}{dataframe of matches between gauges and planning units.
If such a dataframe is not passed (the default, \code{gaugeplanmatch = NULL}),
the function attempts to find a dataframe in \code{dflist} that has both and
create the matching internally. \code{gaugeplanmatch} is used to check
consistency if both \code{gaugefilter} and \code{pufilter} are passed, and to get the
other when only one is passed}

\item{extrasave}{Any other columns to retain from the original datasets that
might be desired attributes later. Default \code{NULL} retains none.}
}
\value{
a tibble with columns for \code{gauge} and \code{PlanningUnitID} (where
included), \code{from} and \code{to} columns indicating the directionality of
pairings, \code{fromtype} and \code{totype} for the type of node being connected, and
an \code{edgeorder} column for the order in which the edges were passed
(previously used for plotting, now deprecated).
}
\description{
Takes a list of causal relationships tables and creates all pairwise
relationships between each of the columns specified in the from-to pairs. Can
be geographically filtered by gauge or planning unit
}
