% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_run_save_ewrs.R
\name{make_ewr_consistent}
\alias{make_ewr_consistent}
\title{handle some different naming conventions}
\usage{
make_ewr_consistent(typearg)
}
\arguments{
\item{typearg}{list of characters for the EWR types}
}
\value{
list of EWR types with standard names
}
\description{
handle some different naming conventions
}
