% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ewrformat.R
\name{make_ewrformat}
\alias{make_ewrformat}
\title{Convert pulled gauge data into the EWR 'Standard time-series' format}
\usage{
make_ewrformat(gaugedata)
}
\arguments{
\item{gaugedata}{a dataframe of gauge data pulled from states or BOM}
}
\value{
gauge data in 'Standard time-series' format
}
\description{
This adds the _flow and _level suffixes and cleans things up
Expects data from either states or BOM, only tested using \link{galenholt/hydrogauge}
If gauges are different lengths or need to be otherwise joined, that should be done afterwards
}
