% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory_setup.R
\name{make_output_dir}
\alias{make_output_dir}
\title{Set up the output directory structure based on parent directory and module at \code{parent_dir/module_output/MODULE_NAME/subdir}}
\usage{
make_output_dir(
  parent_dir,
  scenarios,
  module_name = "EWR",
  subdir = "",
  ewr_outtypes = c("summary", "yearly")
)
}
\arguments{
\item{parent_dir}{parent directory- typically for a project, but could be a single scenario}

\item{scenarios}{names of the scenarios (character vector)}

\item{module_name}{default 'EWR', sets up different directories for different modules}

\item{subdir}{character, default "" for none, but if included, makes a sub-directory in parent_dir/module_output/MODULE_NAME}

\item{ewr_outtypes}{character vector, names of EWR outputs to return. Options include
\itemize{
\item 'summary': outputs summarised to the period
\item 'yearly': outputs summarised to year
\item 'all_events'
\item 'all_successful_events'
\item 'all_interEvents': currently fails in EWR tool, and so not supported
\item 'all_successful_interEvents'
}}
}
\value{
path to output directory
}
\description{
Set up the output directory structure based on parent directory and module at \code{parent_dir/module_output/MODULE_NAME/subdir}
}
