% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plot_table.R
\name{make_plot_table}
\alias{make_plot_table}
\title{Make a table of the aggregations from a plot}
\usage{
make_plot_table(werp_plot)
}
\arguments{
\item{werp_plot}{A ggplot object with WERP-style data underneath. Needs to have the aggregation sequence in columns, not the column name.}
}
\value{
a table with the aggregation step, the function used, and the level of aggregation
}
\description{
Make a table of the aggregations from a plot
}
