% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadpys.R
\name{make_scenario_info_R}
\alias{make_scenario_info_R}
\title{Wrapper for make_scenario_info}
\usage{
make_scenario_info_R(...)
}
\arguments{
\item{...}{arguments for the py function}
}
\value{
file locations as a dict
}
\description{
Wrapper for make_scenario_info
}
