% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_agg_funs.R
\name{Max}
\alias{Max}
\title{Sum aggregation with na.rm = TRUE by default \emph{except} that all-NA vectors still return NA, not 0.}
\usage{
Max(x, na.rm = TRUE)
}
\arguments{
\item{na.rm}{}
}
\description{
Sum aggregation with na.rm = TRUE by default \emph{except} that all-NA vectors still return NA, not 0.
}
