% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_ewr_output.R
\name{maxInterevent}
\alias{maxInterevent}
\title{Get maximum interevent duration of a vector}
\usage{
maxInterevent(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
numeric scalar of the maximum run of 0s
}
\description{
Uses \code{\link[base:rle]{base::rle()}} to get the maximum length of 0s
}
