% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_module_output.R
\name{nameclean}
\alias{nameclean}
\title{Helper to sort the names}
\usage{
nameclean(charvec)
}
\arguments{
\item{charvec}{vector of names}
}
\value{
a character vector
}
\description{
Helper to sort the names
}
