% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_plot_atts.R
\name{node_plot_atts}
\alias{node_plot_atts}
\title{Attach attribute columns to node df to define plot}
\usage{
node_plot_atts(nodedf)
}
\arguments{
\item{nodedf}{a tibble, the dataframe of nodes}
}
\value{
a tibble of nodes with new attribute columns needed to control the
look in \code{DiagrammeR}
}
\description{
This function primarily sets attributes related to size and position.
(fontsize, width, height, x, y) as well as shape and tooltips. It effectively
sets a default look for the structure of the graph. It also needs work- the
numbers are hardcoded, and need to be made relative. It's possible to add
other attributes (see \code{DiagrammeR} documentation), and we'll likely also want
to set some of these (especially height and width) functionally (like we do
with color)
}
\keyword{internal}
