% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation_helpers.R
\name{parse_aggnum}
\alias{parse_aggnum}
\title{Main function of parse_group_until, finds the index for different ways of
specifying group_until}
\usage{
parse_aggnum(x, aggsequence)
}
\arguments{
\item{x}{one of the group_until items}

\item{aggsequence}{as in \code{\link[=multi_aggregate]{multi_aggregate()}}}
}
\value{
numeric index for each group_until
}
\description{
Main function of parse_group_until, finds the index for different ways of
specifying group_until
}
