% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation_helpers.R
\name{parse_char_funs}
\alias{parse_char_funs}
\title{Get an all-character version of the funsequence by parsing entries that are
lists or quosures to characters}
\usage{
parse_char_funs(x)
}
\arguments{
\item{x}{entry in funsequence, typically character, list, or quosure}
}
\value{
character vector
}
\description{
Intended to be used with \code{\link[purrr:map]{purrr::map()}}.
}
