% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation_helpers.R
\name{parse_geo}
\alias{parse_geo}
\title{\code{get} character names for geographic data, allowing aggsequences that are
only characters.}
\usage{
parse_geo(x)
}
\arguments{
\item{x}{character vector or dataframe, typically one entry in the aggregation list}
}
\value{
aggregation sequence, with length-1 character vectors attempted to \code{get}
}
\description{
Only length-1 entries in aggregation sequence are turned into dataframes.
See \code{\link[=multi_aggregate]{multi_aggregate()}} for use. Intended to be used with \code{\link[purrr:map]{purrr::map()}} or similar.
}
\examples{
purrr::map(list(ewr_code = c('ewr_code_timing', 'ewr_code'), sdl_units = "sdl_units"), parse_geo)
}
