% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation_helpers.R
\name{parse_group_until}
\alias{parse_group_until}
\title{Parses group_until into standard form}
\usage{
parse_group_until(group_until, groupers, aggsequence)
}
\arguments{
\item{group_until}{named list of groupers and the step to which they should
be retained. Default NA (retain all groupers for all steps). \emph{FOR EWR USE,
best option is} \code{group_until = list(planning_unit_name = is_notpoint, gauge = is_notpoint)}. This groups by planning unit and gauge until larger
spatial grouping has happened. Leaving 'gauge' off is mathematically safe,
since the gauge geometry forces that grouping, but then the 'gauge' column
gets dropped. Step can be an index, name, or a function that evaluates to
TRUE or FALSE when run on the aggregation sequence. Named list does not
need to contain all groupers, but if so, those that persist throughout
should be given NA or numeric values longer than aggsequence. Vectors the
length of groupers usually work, but are less-well supported.}

\item{groupers}{as in \code{\link[=general_aggregate]{general_aggregate()}}, with the note that these should
be all grouping columns \emph{except} theme and spatial groupings. These are
both automatically added to \code{groupers} according to \code{aggsequence} before
passing to \code{\link[=general_aggregate]{general_aggregate()}}.}

\item{aggsequence}{a named list of aggregation steps in the order to apply
them. Entries for theme aggregation should be character vectors- e.g. \code{name = c('from_theme', 'to_theme')}. Entries for spatial aggregation should be
the sf polygon to aggregate to, e.g. \code{name = sfpolygons} or a length-1
character, e.g. \code{name = "sfpolygons"}. The latter requires the object to be
available with \code{get("sfpolygons")}, but allows passing characters rather
than objects. Not requiring names and is high on the list of improvements.
If we want to be able to re-run from auto-saved metadata params, we need
the names of the spatial levels to match the object, e.g. basin: basin.}
}
\value{
a named list of the group_until groupers with numeric index.
}
\description{
Parses group_until into standard form
}
