% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_types.R
\name{plot_bar}
\alias{plot_bar}
\title{Make a bar plot (qualitative x, usually)}
\usage{
plot_bar(prepped, x_col, x_lab, outcome_lab, position, transoutcome)
}
\arguments{
\item{prepped}{prepared data from \code{plot_data_prep() |>  plot_style_prep()}}

\item{x_col}{character, name of column for the x-axis. Default \code{'scenario'}}

\item{x_lab}{character, default \code{x_col}, allows changing x-label}

\item{outcome_lab}{character, default \code{outcome_col}, allows changing label of
the outcome_col (y-axis or color/fill)}

\item{position}{character or \code{position} function, \code{position} arguments from
\code{\link[ggplot2:geom_bar]{ggplot2::geom_col()}} and \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}} (depending on plot type),
to change from stacked to dodged bars or jitter points. Can be character,
e.g. 'jitter' or a function, e.g. \code{ggplot2::position_jitter(width = 0.1, height = 0)}}

\item{transoutcome}{transformation for outcome as in
\code{\link[ggplot2:scale_continuous]{ggplot2::scale_y_continuous()}} or \code{\link[ggplot2:scale_colour_continuous]{ggplot2::scale_fill_continuous()}}.
Default \code{transoutcome = 'identity'} just uses the data. Most common change
likely \verb{transoutcome = 'log10}}
}
\value{
ggplot object
}
\description{
main function to make a standard bar plot with \code{\link[ggplot2:geom_bar]{ggplot2::geom_col()}}
}
