% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{plot_data_prep}
\alias{plot_data_prep}
\title{Some standard data preparation for plotting}
\usage{
plot_data_prep(
  data,
  outcome_col,
  sceneorder = NULL,
  base_list = NULL,
  zero_adjust = 0,
  onlyzeros = FALSE
)
}
\arguments{
\item{data}{dataframe to prep}

\item{outcome_col}{character, column name for what's plotted on the y-axis.}

\item{sceneorder}{character or factor giving the order to present scenario
levels}

\item{base_list}{NULL (default) or list of arguments for \code{\link[=baseline_compare]{baseline_compare()}}}

\item{zero_adjust}{numeric (default 0) or \code{"auto"}, adjustment to data to
avoid zeros by adding \code{zero_adjust} to \code{abs(data)}, e.g shifting all data
away from zero, either positively or negatively. Zeros themselves are
shifted up or down randomly. Used for avoiding x/0, NaN, and Inf when
relativiszing and taking logs, primarily. Auto shifts by
\code{0.1*min(abs(data[data != 0]))}.}

\item{onlyzeros}{logical, default \code{FALSE}. Should all values be adjusted away from zero (\code{TRUE}) or only adjust zero values (\code{FALSE})?}
}
\value{
a list with prepped versions of \code{data}, \code{outcome_col}, \code{ylab_append} to
be used in plot calls
}
\description{
This is an attempt to pull a lot of copy-paste out of the top of plot
functions. It's likely to continue to evolve quite a bit as we see all the
different things we need to manage in the plot setup. One of the goals here
is to avoid making a million very similar datasets- doing it in functions
keeps those changes sandboxed
}
