% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_types.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Make a heatmap}
\usage{
plot_heatmap(
  prepped,
  x_col,
  x_lab,
  y_col,
  y_lab,
  outcome_lab,
  transoutcome,
  transx,
  transy,
  contour_arglist,
  xtype,
  ytype,
  setLimits,
  base_list
)
}
\arguments{
\item{prepped}{prepared data from \code{plot_data_prep() |>  plot_style_prep()}}

\item{x_col}{character, name of column for the x-axis. Default \code{'scenario'}}

\item{x_lab}{character, default \code{x_col}, allows changing x-label}

\item{y_col}{character, default \code{outcome_col}. Allows separately specifying
y-axis from outcome (e.g. maps, heatmaps, networks)}

\item{y_lab}{character, default y_col, allows changing y-label}

\item{outcome_lab}{character, default \code{outcome_col}, allows changing label of
the outcome_col (y-axis or color/fill)}

\item{transoutcome}{transformation for outcome as in
\code{\link[ggplot2:scale_continuous]{ggplot2::scale_y_continuous()}} or \code{\link[ggplot2:scale_colour_continuous]{ggplot2::scale_fill_continuous()}}.
Default \code{transoutcome = 'identity'} just uses the data. Most common change
likely \verb{transoutcome = 'log10}}

\item{transx}{transformation for x axis as in
\code{\link[ggplot2:scale_continuous]{ggplot2::scale_x_continuous()}}. Default \code{transx = 'identity'} just uses
the data. Most common change likely \verb{transx = 'log10}.}

\item{transy}{transformation of y-axis, \emph{if the outcome is not y}. Default
\code{'identity'}. Ignored if outcome_col is on the y-axis}

\item{contour_arglist}{default \code{NULL}
\itemize{
\item NULL builds a heatmap with \code{\link[ggplot2:geom_tile]{ggplot2::geom_tile()}},
\item \code{list(interpolate = TRUE)} builds an interpolated heatmap with \link[ggplot2:geom_raster(interpolate = TRUE)]{ggplot2::geom_raster(interpolate = TRUE)}
\item \code{list()} builds a \code{\link[ggplot2:geom_contour]{ggplot2::geom_contour()}} with all defaults
\item a named list with names other than 'interpolate' passes those as arguments to \code{\link[ggplot2:geom_contour]{ggplot2::geom_contour()}}
}}

\item{xtype}{character, 'qual', 'numeric' or 'date' (which
covers several different date types). Determines how axes are handled}

\item{ytype}{character, 'qual', 'numeric' or 'date' (which
covers several different date types). Determines how axes are handled}

\item{setLimits}{sets user-supplied color/fill limits for maps, heatmaps, and
networks, or y limits for other plots. Also sets \code{underlay} and \code{overlay}
limits for consistency.}

\item{base_list}{NULL (default) or named list of arguments to
\code{\link[=baseline_compare]{baseline_compare()}};
\itemize{
\item base_lev
\item comp_fun
\item group_cols
\code{\link[=plot_prep]{plot_prep()}} handles \code{zero_adjust}, and other arguments are inferred or
not supported
}}
}
\description{
Make a heatmap
}
