% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hydrographs.R
\name{plot_hydrographs}
\alias{plot_hydrographs}
\title{Standard hydrograph plot}
\usage{
plot_hydrographs(
  hydrolong,
  y_col = "flow",
  gaugefilter = NULL,
  scenariofilter = NULL,
  colors = "RColorBrewer::Dark2",
  sceneorder = NULL,
  scales = "fixed",
  transy = "identity",
  base_lev = NULL,
  comp_fun = NULL,
  ...
)
}
\arguments{
\item{hydrolong}{hydrograph data in long format}

\item{y_col}{character, name of column for the y-axis}

\item{gaugefilter}{set of gauges to filter}

\item{scenariofilter}{set of scenarios to plot, default \code{NULL} plots all of
them}

\item{colors}{name of palette. Will change in future to match plot_outcomes color defining}

\item{sceneorder}{character or factor giving the order to present scenario
levels}

\item{scales}{facet scales, as in \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}. Default \code{scales = 'fixed'} holds them the same, most common change will be to \code{scales = 'free_y'} if gauges have very different flows.}

\item{transy}{transformation for y axis as in \code{\link[ggplot2:scale_continuous]{ggplot2::scale_y_continuous()}}.
Default \code{transy = 'identity'} just uses the data. Most common change likely
\verb{transy = 'log10}}

\item{base_lev}{value to use as the base for comparison. Default NULL, no
comparison. See \code{\link[=baseline_compare]{baseline_compare()}} and \code{\link[=create_base]{create_base()}} for options.}

\item{comp_fun}{function to use in comparison. Default NULL, no comparison.
See \code{\link[=baseline_compare]{baseline_compare()}} and \code{\link[=create_base]{create_base()}} for options.}

\item{...}{passed to \code{\link[=plot_prep]{plot_prep()}}}
}
\value{
ggplot object
}
\description{
Standard hydrograph plot
}
