% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_types.R
\name{plot_numeric}
\alias{plot_numeric}
\title{Make a plot with a numeric x}
\usage{
plot_numeric(
  prepped,
  x_col,
  x_lab,
  outcome_lab,
  position,
  transoutcome,
  transx,
  smooth_arglist,
  xtype = "numeric"
)
}
\arguments{
\item{prepped}{prepared data from \code{plot_data_prep() |>  plot_style_prep()}}

\item{x_col}{character, name of column for the x-axis. Default \code{'scenario'}}

\item{x_lab}{character, default \code{x_col}, allows changing x-label}

\item{outcome_lab}{character, default \code{outcome_col}, allows changing label of
the outcome_col (y-axis or color/fill)}

\item{position}{character or \code{position} function, \code{position} arguments from
\code{\link[ggplot2:geom_bar]{ggplot2::geom_col()}} and \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}} (depending on plot type),
to change from stacked to dodged bars or jitter points. Can be character,
e.g. 'jitter' or a function, e.g. \code{ggplot2::position_jitter(width = 0.1, height = 0)}}

\item{transoutcome}{transformation for outcome as in
\code{\link[ggplot2:scale_continuous]{ggplot2::scale_y_continuous()}} or \code{\link[ggplot2:scale_colour_continuous]{ggplot2::scale_fill_continuous()}}.
Default \code{transoutcome = 'identity'} just uses the data. Most common change
likely \verb{transoutcome = 'log10}}

\item{transx}{transformation for x axis as in
\code{\link[ggplot2:scale_continuous]{ggplot2::scale_x_continuous()}}. Default \code{transx = 'identity'} just uses
the data. Most common change likely \verb{transx = 'log10}.}

\item{smooth_arglist}{NULL (default) or limited list of arguments to
\code{\link[ggplot2:geom_smooth]{ggplot2::geom_smooth()}}. If NULL and x is quantitative, defaults to
straight lines. If just \code{list()}, it uses the defaults in
\code{\link[ggplot2:geom_smooth]{ggplot2::geom_smooth()}}. Available arguments:
\itemize{
\item method
\item method.args
\item se
\item linewidth
\item alpha
If others are desired, we can develop something more general.
}}

\item{xtype}{character, if we're here, should be 'numeric' or 'date' (which
covers several different date types). Dates can't be \code{trans-ed}, so need to
be identified.}
}
\value{
ggplot object
}
\description{
Often (but not always) this is used when scenario is defined numerically.
Uses \code{\link[ggplot2:geom_path]{ggplot2::geom_line()}} or \code{\link[ggplot2:geom_smooth]{ggplot2::geom_smooth()}}. Takes a
\code{smooth_arglist} argument to control that smoother
}
\keyword{internal}
