% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_outcomes.R
\name{plot_outcomes}
\alias{plot_outcomes}
\title{Plot stacked bar graphs of outcomes across scenarios}
\usage{
plot_outcomes(
  outdf,
  outcome_col,
  outcome_lab = outcome_col,
  y_col = outcome_col,
  y_lab = y_col,
  x_col = "scenario",
  x_lab = x_col,
  colorset = "scenario",
  pal_list = "scico::berlin",
  colorgroups = NULL,
  color_lab = ifelse(is.null(colorgroups), colorset, colorgroups),
  plot_type = "2d",
  facet_row = NULL,
  facet_col = NULL,
  facet_wrapper = NULL,
  point_group = NULL,
  sceneorder = NULL,
  scales = "fixed",
  transoutcome = "identity",
  transx = "identity",
  zero_adjust = 0,
  position = "stack",
  base_list = NULL,
  smooth_arglist = NULL,
  underlay_list = NULL,
  overlay_list = NULL,
  setLimits = NULL
)
}
\arguments{
\item{outdf}{dataframe of outcomes, needs a \code{scenario} column}

\item{outcome_col}{character, name of outcome column. This is the data of
interest being plotted. All data operations in \code{\link[=plot_data_prep]{plot_data_prep()}} happen to
this column, and \code{\link[=plot_style_prep]{plot_style_prep()}} bases styling on it. For 2d plots, it is
the y-axis, and for maps, heatmaps, and network plots, it is color/fill. It
is entirely possible to also use it as color/fill for 2d plots in addition
to the y-axis.}

\item{outcome_lab}{character, default \code{outcome_col}, allows changing label of the outcome_col (y-axis or color/fill)}

\item{y_col}{character, default \code{outcome_col}. Allows separately specifying
y-axis from outcome (e.g. maps, heatmaps, networks)}

\item{y_lab}{character, default y_col, allows changing y-lable}

\item{x_col}{character, name of column for the x-axis. Default \code{'scenario'}}

\item{x_lab}{character, default \code{x_col}, allows changing x-label}

\item{colorset}{character, column name defining the colors to use (if
\code{scene_x = TRUE}) or the x-levels (if \code{scene_x = FALSE})- needs to be
renamed to reflect that it isn't just colors}

\item{pal_list}{list of palettes for defining colors for \code{colorset}. Should
be length of \code{colorgroups}}

\item{colorgroups}{character, column name for grouping \code{colorset},
particularly into different palettes}

\item{color_lab}{character, default either \code{colorgroups} or \code{colorset} (if
\code{is.null(colorgroups)}). Allows changing the color legend label}

\item{facet_row}{NULL (default) or character for facet row. Can be \code{'.'} to
have one row and column-facets}

\item{facet_col}{NULL (default) or character for facet column. Can be \code{'.'}
to have one column and row-facets}

\item{facet_wrapper}{NULL (default) or character for column to use for
\code{facet_wrap}}

\item{point_group}{character, column to add additional point groupings to,
e.g. if plotting color by an environmental group, but need separate
lines/points for each environmental objective.}

\item{sceneorder}{character or factor giving the order to present scenario
levels}

\item{scales}{facet scales, as in \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}. Default \code{scales = 'fixed'} holds them the same, most common change will be to \code{scales = 'free_y'} if gauges have very different flows.}

\item{transoutcome}{transformation for y axis as in
\code{\link[ggplot2:scale_continuous]{ggplot2::scale_y_continuous()}}. Default \code{transoutcome = 'identity'} just uses
the data. Most common change likely \verb{transoutcome = 'log10}}

\item{transx}{transformation for x axis as in
\code{\link[ggplot2:scale_continuous]{ggplot2::scale_x_continuous()}}. Default \code{transx = 'identity'} just uses
the data. Most common change likely \verb{transx = 'log10}}

\item{position}{character or \code{position} function, \code{position} arguments from
\code{\link[ggplot2:geom_bar]{ggplot2::geom_col()}} and \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}} (depending on plot type),
to change from stacked to dodged bars or jitter points. Can be character,
e.g. 'jitter' or a function, e.g. \code{ggplot2::position_jitter(width = 0.1, height = 0)}}

\item{base_list}{NULL (default) or limited list of arguments to
\code{\link[=baseline_compare]{baseline_compare()}};
\itemize{
\item base_lev
\item comp_fun
\item group_cols
\code{\link[=plot_prep]{plot_prep()}} handles \code{zero_adjust}, and other arguments are inferred or
not supported
}}

\item{smooth_arglist}{NULL (default) or limited list of arguments to
\code{\link[ggplot2:geom_smooth]{ggplot2::geom_smooth()}}. If NULL and x is quantitative, defaults to
straight lines. Arguments
\itemize{
\item method
\item method.args
\item se
\item linewidth
If others are desired, we can develop something more general.
}}

\item{underlay_list}{default NULL, named list (or list of named lists for
multiple underlay levels) of arguments to plot a map underlying the main
map data. Names define arguments, \code{underlay} is required, either character
or an sf, \code{underlay_pal} do define colors, can be single color or paletteer
name, and \code{underlay_ycol} if coloring the underlay by values. Multiple
values (e.g. having an \code{underlay_ycol} and a palette for \code{underlay_pal})
only works if the main data is not the same type- we can't use different
palettes for underlay fill and main data fill, for example, but can if the
underlay is fill (polygons) and the main data is points.}

\item{overlay_list}{as \code{underlay_list}, but names \code{"overlay_*"}}

\item{setLimits}{sets user-supplied color/fill limits for maps, heatmaps, and networks, or y limits
for other plots. Also sets \code{underlay} and \code{overlay} limits for consistency.}

\item{smooth_method}{\code{method} argument to \code{\link[ggplot2:geom_smooth]{ggplot2::geom_smooth()}}. Ignored
if not smoothing}

\item{smooth_args}{\code{method.args} argument to \code{\link[ggplot2:geom_smooth]{ggplot2::stat_smooth()}}.
Ignored if not smoothing.}

\item{smooth_se}{\code{se} argument to \code{\link[ggplot2:geom_smooth]{ggplot2::geom_smooth()}}. Ignored if not
smoothing}
}
\value{
a ggplot stacked bar plot with standard formatting and data cleaning
}
\description{
Has several ways to color, including grouped palettes. Lets us choose whether
to stack scenarios with outcome groups on x (\code{scene_x = FALSE}) or stack
outcome groups with scenarios on x (\code{scene_x = TRUE}). I do \emph{not} provide an
argument to drop legend labels on purpose- it's dangerous to auto-drop them.
The user can always \code{+ theme(legend.position = 'none')} to the returned
object.
}
