% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{plot_style_prep}
\alias{plot_style_prep}
\title{Title}
\usage{
plot_style_prep(
  prepped,
  colorset,
  colorgroups,
  pal_list,
  transoutcome,
  transx,
  point_group
)
}
\arguments{
\item{point_group}{}
}
\description{
Title
}
