% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_ewr_output.R
\name{prep_ewr_output}
\alias{prep_ewr_output}
\title{Make EWR results with achievement for ongoing use}
\usage{
prep_ewr_output(
  dat,
  type = "achievement",
  year_roll = "best",
  gaugefilter = NULL,
  scenariofilter = NULL,
  add_max = FALSE
)
}
\arguments{
\item{dat}{EWR output dataframe (usually for all gauges and scenarios)}

\item{type}{as in \code{\link[=read_and_geo]{read_and_geo()}}, but with two special options:
\itemize{
\item 'achievement', calculates EWR achievement from 'yearly',
\item 'interevents', calculates some interevent values from all_interEvents or all_successful_interEvents
For the EWR tool, the direct options are
\item 'summary',
\item 'yearly',
\item 'all_events',
\item 'all_successful_events',
\item 'all_interEvents',
\item 'all_successful_interEvents'
}}

\item{year_roll}{character 'best' or number, specific number of years to check assessment for. 'Best' uses 10-year windows if possible. 1 uses the NSW method.}

\item{gaugefilter}{subset of gauges, default NULL}

\item{scenariofilter}{subset of scenarios, default NULL}

\item{add_max}{logical, default FALSE. Add a 'MAX' scenario that passes all EWRs, usable as a reference, but often causes problems. Usually better to just calculate it post-hoc}
}
\value{
a tibble with ewr_achieved
}
\description{
Make EWR results with achievement for ongoing use
}
