% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_run_save_ewrs.R
\name{prep_run_save_ewrs_old}
\alias{prep_run_save_ewrs_old}
\title{Set up, run, and (possibly) save EWR outputs}
\usage{
prep_run_save_ewrs_old(
  hydro_dir,
  output_parent_dir,
  scenarios = NULL,
  model_format = "Standard time-series",
  climate = "Standard - 1911 to 2018 climate categorisation",
  outputType = "none",
  returnType = "none",
  MINT = (100 - 0)/100,
  MAXT = (100 + 0)/100,
  DUR = (100 - 0)/100,
  DRAW = (100 - 0)/100,
  scenario_filename_split = "_DIRECTORYAPPEND_",
  extrameta = NULL,
  datesuffix = FALSE,
  rparallel = FALSE
)
}
\arguments{
\item{hydro_dir}{Directory containing hydrographs. Can be an outer directory,
e.g. \code{hydrographs} that splits into scenario subdirs, or can be a single
scenario subdir.}

\item{output_parent_dir}{parent directory for the outputs. Can be anything,
but two typical cases:
\itemize{
\item The directory containing \code{hydro_dir}, which puts the \code{module_outputs} at the same level as the hydrographs
\item If running in batches for single scenarios, may be \code{hydro_dir}, which just puts the \code{module_outputs} in \code{hydro_dir}
}}

\item{scenarios}{\code{NULL} (default) or character.
\itemize{
\item \code{NULL}- finds scenario names by parsing directory names in \code{hydro_dir}. If no internal directories, just stays in \code{hydro_dir}. This captures the two typical situations discussed for \code{output_parent_dir}. If there are other directories in \code{hydro_dir} that do not contain hydrological scenarios, should use a character vector.
\item character vector of scenario names. This allows unusual directory structures.
}}

\item{model_format}{see EWR tool. One of:
\itemize{
\item 'IQQM - NSW 10,000 years' (default)
\item 'Bigmod - MDBA'
\item 'Source - NSW (res.csv)'
}}

\item{climate}{see EWR tool. One of:
\itemize{
\item 'Standard - 1911 to 2018 climate categorisation' (default, with categories)
\item 'NSW 10,000 year climate sequence' (not categorised, just a blank csv for 10,000 years that allows it to run)
}}

\item{outputType}{list of strings or character vector defining what to save
to disk. One or more of:
\itemize{
\item 'none' (default), do not save outputs- ignored if in a list with others
\item 'summary',
\item 'yearly',
\item 'all_events',
\item 'all_successful_events',
\item 'all_interEvents', # Does not work with current EWR tool
\item 'all_successful_interEvents'
}}

\item{returnType}{list of strings or character vector defining what to return
to the active R session. Same options as \code{outputType}}

\item{MINT}{see EWR}

\item{MAXT}{see EWR}

\item{DUR}{see EWR}

\item{DRAW}{see EWR}

\item{scenario_filename_split}{character (including regex) to split the scenario filenames in the 'scenario' column of the EWR outputs. Mostly for handling the situation of gauge-named csvs with the scenario name appended to the front. When auto-appended, '\emph{DIRECTORYAPPEND}' is used. If the appending has been done by the user, will need to pass in the split. The 'scenario' is given the first piece of the split, so do not use a split pattern in the scenario name, e.g. do not name scenarios a_1 and then append csvs with "_" like a_1_401234.csv, or the _1 will be lost..}

\item{extrameta}{list, extra information to include in saved metadata documentation for the run. Default NULL.}

\item{datesuffix}{logical. whether to add a suffix to saved filenames to
provide a datestamp. Should be deprecated in favour of metadata files.}

\item{rparallel}{logical, default FALSE. If TRUE, parallelises over the scenarios in hydro_dir using \code{furrr}. To use, install \code{furrr} and set a \code{\link[future:plan]{future::plan()}} (likely \code{multisession} or \code{multicore})}
}
\value{
a list of dataframe(s) if \code{returnType} is not 'none', otherwise, NULL
}
\description{
This does some directory setup and parsing, runs the EWR tool and, if asked,
saves the output. If the output saves, it also auto-saves both yaml and json
metadata files with all parameters needed to run this part of the toolkit
with parameters. Scenario metadata is prepended, if found.
}
