% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_and_geo.R
\name{read_and_geo}
\alias{read_and_geo}
\title{Prepare ewr results from the EWR tool for use in aggregation}
\usage{
read_and_geo(datapath, type = "achievement", geopath, whichcrs = 4283, ...)
}
\arguments{
\item{datapath}{path to module results (Often EWR outputs with HydroBOT
modifications to format)}

\item{type}{character, a grep for the files to choose. Two special cases:
\itemize{
\item 'achievement', calculates EWR achievement from 'yearly',
\item 'everything', gets all files
For the EWR tool, the direct options are
\item 'summary',
\item 'yearly',
\item 'all_events',
\item 'all_successful_events',
\item 'all_interEvents',
\item 'all_successful_interEvents'
}}

\item{geopath}{sf object with geographic locations matching a column in the data, or path to a csv with gauge locations in lat/long (assumes BOM
currently) or a shapefile}

\item{whichcrs}{desired crs for the output}

\item{...}{passes \code{gaugefilter}, \code{scenariofilter}, and \code{add_max} to \code{\link[=prep_ewr_output]{prep_ewr_output()}}.
Particularly useful for only processing some of the data at a time (e.g. in
parallel)}
}
\value{
an \code{sf} dataframe with the desired outputs including scenario
names and gauge or polygon locations as a POINT, POLYGON, or MULTIPOLYGON geometry. Primary purpose is to
prepare data to send to \code{\link[=multi_aggregate]{multi_aggregate()}}
}
\description{
This reads in the EWR results given a path and \code{type} (summary or annual),
and pairs them to their geographic information
}
