% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hydro.R
\name{read_hydro}
\alias{read_hydro}
\title{Wrapper to read hydrographs into R}
\usage{
read_hydro(
  hydropath,
  scenariofilter = NULL,
  long = TRUE,
  format = "csv",
  gaugemap = "iqqm"
)
}
\arguments{
\item{hydropath}{path to the hydrographs}

\item{scenariofilter}{character vector of scenario names to include. Default \code{NULL} includes all}

\item{long}{logical, default TRUE. Do we want the data long (with a column for gauge number) or wide (FALSE, with gauge numbers as columns)}

\item{format}{character, default 'csv'. Format of the hydrographs, determines which reading function gets called.}
}
\value{
tibble of hydrographs
}
\description{
Wrapper to read hydrographs into R
}
