% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hydro.R
\name{read_hydro_csv}
\alias{read_hydro_csv}
\title{Read csv hydrographs in the standard directory structure}
\usage{
read_hydro_csv(hydropath, long)
}
\arguments{
\item{hydropath}{path to the hydrographs}

\item{long}{logical, default TRUE. Do we want the data long (with a column for gauge number) or wide (FALSE, with gauge numbers as columns)}
}
\value{
tibble of hydrographs
}
\description{
Standard structure is
hydropath(name of set of scenarios) ->
hydrographs ->
dirs for scenario names ->
csv files named SCENARIO_NAME.csv, with cols for date and each gauge
This last bit could have files for each gauge separately (was the original format)
}
