% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hydro.R
\name{read_hydro_nc}
\alias{read_hydro_nc}
\title{Read netcdf hydrographs in the standard directory structure}
\usage{
read_hydro_nc(hydropath, scenariofilter, long, gaugemap)
}
\arguments{
\item{hydropath}{path to the hydrographs}

\item{scenariofilter}{character vector of scenario names to include. Default \code{NULL} includes all}

\item{long}{logical, default TRUE. Do we want the data long (with a column for gauge number) or wide (FALSE, with gauge numbers as columns)}

\item{gaugemap}{mapping of gauges from nodes. Default 'iqqm' gets it from EWR tool, otherwise, a dataframe with columns 'node' and 'gauge'}
}
\value{
tibble of hydrographs
}
\description{
Uses the machinery of \link{find_scenario_paths} to get paths and names of the
hydrographs, read them in, and name them.
}
