% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_comparison_funs.R
\name{relative}
\alias{relative}
\title{Division (relative comparison)}
\usage{
relative(x, y, add_eps = 0)
}
\arguments{
\item{x}{data to compare}

\item{y}{reference data}

\item{add_eps}{default 0. epsilon to add to prevent divide-by-zero infs. Can be \code{'auto'} to use half of the minimum nonzero value}
}
\value{
numeric, as \code{/}
}
\description{
Easily-callable wrapper for \code{/}, with ability to avoid divide-by-0
}
