% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{resource_plan_areas}
\alias{resource_plan_areas}
\title{Resource plan areas}
\format{
\subsection{\code{resource_plan_areas} An {sf} object:}{

\describe{
\item{SWWRPANAME, SWWRPACODE}{Name and code for resource plan area}
\item{STATE}{State}
}
}
}
\source{
\url{https://data.gov.au/data/dataset/7b0c274f-7f12-4062-9e54-5b8227ca20c4/resource/8369c483-975c-4f54-8eda-a6fbd0ea89ad/download/sworkingadhocjobsp3827dataoutputsupdate-april-2019sw-wrpasurface-water-water-resource-plan-areas.zip}
}
\usage{
resource_plan_areas
}
\description{
Polygons defining the boundaries of resource plan areas. Sourced from \href{https://www.mdba.gov.au/publications-and-data/maps-and-spatial-data/spatial-data}{MDBA data website},
with direct link given below. Polygons have been slightly simplified (reduced
vertices) to save size.
}
\keyword{datasets}
