% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_ewr_output.R
\name{roll_frequency}
\alias{roll_frequency}
\title{Helper to get the frequency of occurrence without clogging up the mutates}
\usage{
roll_frequency(x, year_roll, pad_initial = FALSE, na.rm = FALSE)
}
\arguments{
\item{x}{vector to calculate rolling frequencies for}

\item{year_roll}{window size for the roll}

\item{pad_initial}{Allow calculating values in the first years < year_roll; roll the year_roll if possible, otherwise as much as possible. Note that this makes the 1:year_roll entries less smoothed.}

\item{na.rm}{default FALSE, na action for the sums over lags- TRUE is tempting, but can cause fails for high thresholds when we don't actually know whether they're met, and so FALSE is more appropriate (keeps intermediates NA).}
}
\value{
numeric vector
}
\description{
Helper to get the frequency of occurrence without clogging up the mutates
}
\keyword{internal}
