% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_ewr_output.R
\name{roll_interevent}
\alias{roll_interevent}
\title{Helper to get the rolling maximum interevent duration without clogging up the mutates}
\usage{
roll_interevent(x, year_roll, pad_initial = FALSE)
}
\arguments{
\item{x}{vector to calculate rolling interevents for}

\item{year_roll}{window size for the roll}

\item{pad_initial}{Allow calculating values in the first years < year_roll; roll the year_roll if possible, otherwise as much as possible. Note that this makes the 1:year_roll entries less smoothed.}
}
\value{
numeric vector
}
\description{
Helper to get the rolling maximum interevent duration without clogging up the mutates
}
\keyword{internal}
