% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadpys.R
\name{run_save_ewrs_R}
\alias{run_save_ewrs_R}
\title{Wrapper for run_save_ewrs}
\usage{
run_save_ewrs_R(...)
}
\arguments{
\item{...}{arguments to the py function}
}
\value{
the EWRs, and saving
}
\description{
Wrapper for run_save_ewrs
}
