% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_runner.R
\name{run_toolkit_params}
\alias{run_toolkit_params}
\title{Run the toolkit through the aggregation step with a file of yaml parameters
or passing in yaml as characters.}
\usage{
run_toolkit_params(
  yamlpath = NULL,
  passed_args = NULL,
  defaults = system.file("yml/default_params.yml", package = "werptoolkitr")
)
}
\arguments{
\item{yamlpath}{character, path to yaml file of parameters. Only needs to
contain values for parameters for which we wish to modify \code{defaults}}

\item{passed_args}{character, in yaml format, e.g. ""scenario_dir:
'project_dir/hydrographs'". Likely possible to pass more than one, but
complicated to get the yaml right. Suggest in that case modifying the
params.yml}

\item{defaults}{character, path to default set of yaml parameters, so the
\code{yamlpath} file and \code{passed_args} only need to include those that are
modified}
}
\value{
runs the toolkit, returns NULL
}
\description{
The plan is for this to make scripting easier, and be able to change single
simple parameters (e.g. directory) from command line or in other workflows
}
