% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory_setup.R
\name{scenario_names_from_hydro}
\alias{scenario_names_from_hydro}
\title{Used to auto-acquire scenario names from the hydrograph directory}
\usage{
scenario_names_from_hydro(hydro_dir)
}
\arguments{
\item{hydro_dir}{path to hydrograph directory. Can be the outer directory with all, or potentially a single scenario}
}
\value{
character vector of scenario names
}
\description{
DEPRECATED, only used for the old prep_run_save_ewrs
}
\keyword{internal}
