% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sdl_units}
\alias{sdl_units}
\title{Sustainable diversion limits units}
\format{
\subsection{\code{sd_units} An {sf} object:}{

\describe{
\item{SWSDLID, SWSDLName}{ID and name of SDL unit}
\item{StateID}{State}
}
}
}
\source{
\url{https://data.gov.au/data/dataset/4afa3227-8557-4bb6-944a-6494b28ae160/resource/988ae947-f4dc-4f75-9773-536df71b03ad/download/sworkingadhocjobsp3827dataoutputsupdate-january-2019surface-water-sdl-resource-units.zip}
}
\usage{
sdl_units
}
\description{
Polygons defining the boundaries of SDL units. Sourced from \href{https://www.mdba.gov.au/publications-and-data/maps-and-spatial-data/spatial-data}{MDBA data website},
with direct link given below. Polygons have been slightly simplified (reduced
vertices) to save size.
}
\keyword{datasets}
