% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_ewr_output.R
\name{separate_ewr_codes}
\alias{separate_ewr_codes}
\title{Parser for EWR codes into the main code and the extra bits (called 'timing' for some reason)}
\usage{
separate_ewr_codes(df)
}
\arguments{
\item{df}{a dataframe with an ewr_code column with raw ewr names (e.g. EWR outputs, causal mappings)}
}
\value{
a tibble with a clean ewr_code_main column and an ewr_code column with the full EWR name
}
\description{
This allows keeping the whole ewr table, the EWR outputs, and causal networks matched and standardized parsing.
}
