% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_agg_funs.R
\name{SpatialWeightedMean}
\alias{SpatialWeightedMean}
\title{Weighted mean by \code{area} column}
\usage{
SpatialWeightedMean(x, na.rm = TRUE)
}
\arguments{
\item{x}{values to aggregate}

\item{na.rm}{default \code{na.rm = TRUE} finds the aggregation ignoring NA, opposite usual na.rm default \code{na.rm = FALSE}, which returns NA if there are any NA.}
}
\value{
summarised data
}
\description{
Finds the arithmetic mean weighted by the \code{area} column, which is auto-created in the spatial datasets.
}
