% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{summary_ewr_output}
\alias{summary_ewr_output}
\title{Example output from EWR tool}
\format{
\subsection{\code{summary_ewr_output} A data frame:}{

\describe{
\item{scenario, gauge}{Name of scenario and gauge at which EWR is assessed}
\item{ewr_code, ewr_code_timing, multigauge}{EWR code and timing suffix, and whether it requires multiple gauges}
\item{event_years, frequency, ...}{Information about the EWR achievement}
}
}
}
\source{
Internal, from running EWR tool on \code{extdata/testsmall}
}
\usage{
summary_ewr_output
}
\description{
Example output of EWR tool using the simple hydrographs provided in
extdata/testsmall. Primarily to allow testing and checking without needing to
run the EWR tool.
}
\keyword{datasets}
