% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{test_overplotting}
\alias{test_overplotting}
\title{Check silent overplotting}
\usage{
test_overplotting(
  data,
  facet_wrapper,
  facet_row,
  facet_col,
  x_col = NULL,
  y_col = NULL
)
}
\arguments{
\item{data}{the prepped data to be plotted}

\item{facet_wrapper}{NULL (default) or character for column to use for
\code{facet_wrap}}

\item{facet_row}{NULL (default) or character for facet row. Can be \code{'.'} to
have one row and column-facets}

\item{facet_col}{NULL (default) or character for facet column. Can be \code{'.'}
to have one column and row-facets}

\item{x_col}{character, name of column for the x-axis. Default \code{'scenario'}}

\item{y_col}{character, default \code{outcome_col}. Allows separately specifying
y-axis from outcome (e.g. maps, heatmaps, networks)}
}
\value{
TRUE if there's no overplotting, otherwise aborts
}
\description{
Fill in maps and heatmaps can silently plot multiple values on top of each
other. THis checks data dimension to ensure only a single value is plotted
}
