% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_ewr_output.R
\name{assess_ewr_interevents}
\alias{assess_ewr_interevents}
\title{Some calculations needed for aggregation of interevent statistics}
\usage{
assess_ewr_interevents(interdf)
}
\arguments{
\item{interdf}{the all_interEvents or all_successful_interEvents EWR output}
}
\value{
tibble with columns \code{scenario}, \code{gauge}, \code{planning_unit_name},
\code{state}, \code{SWSDLName}, \code{ewr_code}, \code{start_date}, \code{inter_event_length},
\code{ewr_code_timing}, \code{max_interevent}, \code{exceedance_days} (realised - max),
\code{interevent_ratio} (realised / max), \code{exceedance_ratio} (interevent_ratio -
1), \code{exceedance} (binary; realised >= max), \code{exceedance_only} (days above
max), \verb{days_in_exceeding (all realised if max passed)}
}
\description{
Some calculations needed for aggregation of interevent statistics
}
