% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_ewr_output.R
\name{assess_ewr_interevents}
\alias{assess_ewr_interevents}
\title{Some calculations needed for aggregation of interevent statistics}
\usage{
assess_ewr_interevents(interdf)
}
\arguments{
\item{interdf}{the all_interEvents EWR output}
}
\value{
dataframe with columns 'scenario', 'gauge', 'planning_unit_name',
'state', 'SWSDLName', 'ewr_code', 'start_date', 'inter_event_length',
'ewr_code_timing', 'max_interevent', 'exceedance_days' (realised - max),
'interevent_ratio' (realised / max), 'exceedance_ratio' (interevent_ratio -
1), 'exceedance' (binary; realised >= max), 'exceedance_only' (days above
max), 'days_in_exceeding (all realised if max passed)'
}
\description{
Some calculations needed for aggregation of interevent statistics
}
