% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{basin_rivers}
\alias{basin_rivers}
\title{River lines in the basin}
\format{
\subsection{\code{basin_rivers} An \code{sf} object:}{

\describe{
\item{geometry}{The geometry, all LINESTRING or MULTILINESTRING}
\item{other values}{As documented at http://www.bom.gov.au/water/geofabric/}
}
}
}
\source{
BOM, \url{http://www.bom.gov.au/water/geofabric/}
}
\usage{
basin_rivers
}
\description{
Spatial lines for the major rivers in the basin, from geofabric. Has been clipped to basin and paths simplified.
}
\keyword{datasets}
