% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_ewr_obj.R
\name{clean_ewr_obj}
\alias{clean_ewr_obj}
\title{Pair EWR indicators to environmental objectives}
\usage{
clean_ewr_obj(ewrobjpath, gaugescale = TRUE, saveout = FALSE, outdir, savename)
}
\arguments{
\item{ewrobjpath}{path to the ewr-obj mapping.}

\item{gaugescale}{logical, whether to return data at the gauge scale by remapping from LTWPShortName to gauge (\code{TRUE}), or to just leave at the LTWPShortName (e.g. sdl unit) scale as NSW now defines the relationships (\code{FALSE})}

\item{saveout}{one of FALSE, 'r', or 'csv', controls saving and type.
\itemize{
\item \code{FALSE}: (the default): Does not save anything
\item \code{'r'}: saves an rdata file, used primarily for building the data directory in a package
\item \code{'csv'}: saves a csv with the name \code{outdir/savename_YearMonthDayHourMinute.csv}
}}

\item{outdir}{path to the directory to save into. not needed if \code{saveout} is \code{FALSE}}

\item{savename}{character for the filename to save. the date and time gets appended to avoid overwriting}
}
\value{
A \code{tibble} with columns for \code{LTWPShortName}, \code{ewr_code}, \code{ewr_code_timing}, and \code{env_obj} (if \code{gaugescale = FALSE}), and including \code{PlanningUnitID} and \code{gauge} if \code{gaugescale = TRUE})
}
\description{
Bespoke function to clean the EWR table and LTWP table, extract environmental objectives, and pair them. Will need to change or at least extensively retest if those datasets change
}
