% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_ewr_output.R
\name{clean_yearly}
\alias{clean_yearly}
\title{Additional cleanup specific to yearly EWR outputs.}
\usage{
clean_yearly(annualdf)
}
\arguments{
\item{annualdf}{the result of \code{\link[=get_module_output]{get_module_output()}} for EWR data with \code{type = 'yearly'}}
}
\value{
tibble of the cleaned yearly EWR output
}
\description{
Additional cleanup specific to yearly EWR outputs.
}
