% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory_setup.R
\name{make_output_dir}
\alias{make_output_dir}
\title{Set up the output directory structure based on parent directory and module}
\usage{
make_output_dir(parent_dir, scenarios, module_name = "EWR")
}
\arguments{
\item{parent_dir}{parent directory- typically for a project, but could be a single scenario}

\item{scenarios}{names of the scenarios (character vector)}

\item{module_name}{default 'EWR', sets up different directories for different modules}
}
\value{
path to output directory
}
\description{
Set up the output directory structure based on parent directory and module
}
