% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{planning_units}
\alias{planning_units}
\title{Planning Units for long-term watering plans}
\format{
\subsection{\code{planning_units} An {sf} object:}{

\describe{
\item{LTWPShortName, PlanningUnitName, geometry}{The long-term watering plan area and planning unit name matching, to the extent possible, the EWR table. Geometry is polygons. Other columns for reference.}
}
}
}
\source{
MDBA Environmental Assets and Functions Database
}
\usage{
planning_units
}
\description{
Polygons defining planning units. Currently just for NSW, sourced from MDBA
Environmental Assets and Functions Database
}
\keyword{datasets}
