% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hydrographs.R
\name{plot_hydrographs}
\alias{plot_hydrographs}
\title{Standard hydrograph plot}
\usage{
plot_hydrographs(
  hydrolong,
  outcome_col = "flow",
  gaugefilter = NULL,
  scenariofilter = NULL,
  colors = "RColorBrewer::Dark2",
  sceneorder = NULL,
  scales = "fixed",
  transoutcome = "identity",
  base_lev = NULL,
  comp_fun = NULL,
  ...
)
}
\arguments{
\item{hydrolong}{hydrograph data in long format}

\item{outcome_col}{character, name of outcome column. This is the data of
interest being plotted. All data operations in \code{\link[=plot_data_prep]{plot_data_prep()}} happen to
this column, and \code{\link[=plot_style_prep]{plot_style_prep()}} bases styling on it. For 2d plots, it is
the y-axis, and for maps, heatmaps, and network plots, it is color/fill. It
is entirely possible to also use it as color/fill for 2d plots in addition
to the y-axis.}

\item{gaugefilter}{set of gauges to filter}

\item{colors}{name of palette. Will change in future to match plot_outcomes color defining}

\item{sceneorder}{character or factor giving the order to present scenario
levels}

\item{scales}{facet scales, as in \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}. Default \code{scales = 'fixed'} holds them the same, most common change will be to \code{scales = 'free_y'} if gauges have very different flows.}

\item{transoutcome}{transformation for y axis as in
\code{\link[ggplot2:scale_continuous]{ggplot2::scale_y_continuous()}}. Default \code{transoutcome = 'identity'} just uses
the data. Most common change likely \verb{transoutcome = 'log10}}

\item{...}{passed to \code{\link[=plot_prep]{plot_prep()}}}
}
\value{
ggplot object
}
\description{
Standard hydrograph plot
}
