% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_ewr_output.R
\name{prep_ewr_output}
\alias{prep_ewr_output}
\title{Make EWR results with achievement for ongoing use}
\usage{
prep_ewr_output(
  dat,
  type = "achievement",
  year_roll = "best",
  gaugefilter = NULL,
  scenariofilter = NULL,
  add_max = TRUE
)
}
\arguments{
\item{type}{character, a grep for the files to choose. Two special cases:
\itemize{
\item 'achievement', calculates EWR achievement from 'yearly',
\item 'everything', gets all files
For the EWR tool, the direct options are
\item 'summary',
\item 'yearly',
\item 'all_events',
\item 'all_successful_events',
\item 'all_interEvents',
\item 'all_successful_interEvents'
}}

\item{year_roll}{character 'best' or number, specific number of years to check assessment for. 'Best' uses 10-year windows if possible. 1 uses the NSW method.}

\item{gaugefilter}{subset of gauges, default NULL}

\item{scenariofilter}{subset of scenarios, default NULL}

\item{add_max}{logical, default TRUE. Add a 'MAX' scenario that passes all EWRs, usable as a reference}

\item{dir}{path to directory with the EWR output for all gauges and scenarios}
}
\value{
a tibble with ewr_achieved
}
\description{
Make EWR results with achievement for ongoing use
}
