% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_and_agg.R
\name{read_and_agg}
\alias{read_and_agg}
\title{Read in data and aggregate along theme and spatial dimension}
\usage{
read_and_agg(
  datpath,
  type,
  geopath,
  causalpath,
  groupers = "scenario",
  group_until = rep(NA, length(groupers)),
  aggCols,
  aggsequence,
  funsequence,
  saveintermediate = FALSE,
  namehistory = TRUE,
  keepAllPolys = FALSE,
  failmissing = TRUE,
  auto_ewr_PU = FALSE,
  pseudo_spatial = NULL,
  returnList = TRUE,
  savepath = NULL,
  extrameta = NULL,
  rparallel = FALSE,
  par_recursive = TRUE,
  savepar = "combine",
  add_max = TRUE,
  ...
)
}
\arguments{
\item{datpath}{path to indicator data, or indicator data itself as a data
frame. Currently needs to be EWR (same as \code{ewrpath} argument in
\code{\link[=read_and_geo]{read_and_geo()}}), but left more general here for future}

\item{type}{character, a grep for the files to choose. Two special cases:
\itemize{
\item 'achievement', calculates EWR achievement from 'yearly',
\item 'everything', gets all files
For the EWR tool, the direct options are
\item 'summary',
\item 'yearly',
\item 'all_events',
\item 'all_successful_events',
\item 'all_interEvents',
\item 'all_successful_interEvents'
}}

\item{geopath}{sf object with geographic locations matching a column in the data, or path to a csv with gauge locations in lat/long (assumes BOM
currently) or a shapefile}

\item{causalpath}{path to the causal relationships .rds file or the causal
network list object or its name}

\item{groupers}{as in \code{\link[=general_aggregate]{general_aggregate()}}, with the note that these should
be all grouping columns \emph{except} theme and spatial groupings. These are
both automatically added to \code{groupers} according to \code{aggsequence} before
passing to \code{\link[=general_aggregate]{general_aggregate()}}.}

\item{group_until}{named list of groupers (column names) and the step to
which they should be retained. Default NA (retain all groupers for all
steps). \emph{FOR EWR USE, best option is} \code{group_until = list(SWSDLName = 'sdl_units', planning_unit_name = 'sdl_units', gauge = is_notpoint)}. This
groups by planning unit and gauge and planning unit until larger spatial
grouping has happened, dealing with the issue of gauges reporting into
multiple PUs and SDLs. Leaving 'gauge' off is mathematically safe, since
the gauge geometry forces that grouping, but then the 'gauge' column gets
dropped. Step can be an index, name, or a function that evaluates to TRUE
or FALSE when run on the aggregation sequence. Named list does not need to
contain all groupers, but if so, those that persist throughout should be
given NA or numeric values longer than aggsequence. Vectors the length of
groupers usually work, but are less-well supported.}

\item{aggCols}{an expression for the columns to aggregate (the data columns).
See \code{selectcreator} for formats}

\item{aggsequence}{a named list of aggregation steps in the order to apply
them. Entries for theme aggregation should be character vectors- e.g. \code{name = c('from_theme', 'to_theme')}. Entries for spatial aggregation should be
the sf polygon to aggregate to, e.g. \code{name = sfpolygons} or a length-1
character, e.g. \code{name = "sfpolygons"}. The latter requires the object to be
available with \code{get("sfpolygons")}, but allows passing characters rather
than objects. Not requiring names and is high on the list of improvements.
If we want to be able to re-run from auto-saved metadata params, we need
the names of the spatial levels to match the object, e.g. basin: basin.}

\item{funsequence}{a list of aggregation functions to apply in the order to
apply them. Each list entry can be one value, e.g. a character or bare
name, or can be multiple if multiple aggregations should be done at that
step, e.g. \code{c('ArithmeticMean', 'LimitingFactor')}. The entries can also be
lists themselves, useful for passing functions with arguments, e.g \code{list(wm = ~weighted.mean(., w = area, na.rm = TRUE))}. \emph{Important:} as of \code{dplyr}
1.1, if these are anonymous functions that refer to data variables (like
the \code{w = area} argument in the \code{\link[=weighted.mean]{weighted.mean()}} example), that list needs
to be wrapped in \code{\link[rlang:defusing-advanced]{rlang::quo()}}, e.g. \verb{rlang::quo(list(wm = ~weighted.mean(., w = area, na.rm = TRUE))}. And we can no longer mix
character and other forms in the same sub-list (single aggregation step).}

\item{saveintermediate}{logical, default \code{FALSE}. * \code{FALSE} (the default):
Save only the final result as a tibble or sf * \code{TRUE}: Save every step of
the aggregation as a tibble or sf in a list}

\item{namehistory}{logical, default \code{TRUE}.
\itemize{
\item \code{TRUE} (the default): The name of the aggregated column(s) retain the full aggregation history of the form
\code{agglevelN_aggfunctionN_...agglevel1_aggfunction1_originalcolumn}. This is
ugly, but saves memory and says exactly what the values in each column are.
\item \code{FALSE}: The aggregation history is moved out of the column names and
into new columns that define it using \code{\link[=agg_names_to_cols]{agg_names_to_cols()}}. The column
name(s) become(s) the original column name(s) specified by \code{aggCols}. This
is far cleaner and easier for analysis (e.g. filtering on aggregation
functions at a particular step), but increases the size of the dataset and
the meaning of the values in the aggregation column have to be interpreted
with the values in the new columns defining history.
}}

\item{keepAllPolys}{logical, default \code{FALSE}. Should polygons in \code{to_geo}
that have no values in \code{dat} be retained? The default \code{FALSE} keeps NA
polygons from cluttering things up, but \code{TRUE} can be useful to not lose
them, especially for later plotting. However, it is typically best from a
data and cleanliness perspective to use \code{FALSE} here and use the bare set
of polys as an \code{underlay} in \code{\link[=plot_outcomes]{plot_outcomes()}}.}

\item{failmissing}{logical, default \code{TRUE}: fail if the requested grouping or
aggregation columns not exist. If \code{FALSE}, proceed with those that do exist
and silently drop those that don't. Similar to \code{tidyselect::all_of()} vs
\code{tidyselect::any_of()} in \code{tidyselect}}

\item{auto_ewr_PU}{logical, default \code{FALSE}. Auto-detect EWRs and enforce
appropriate theme and spatial scaling related to gauges and planning units,
as defined in \code{\link[=theme_aggregate]{theme_aggregate()}} and \code{\link[=spatial_aggregate]{spatial_aggregate()}}. Specifically,
if \code{TRUE}, this automatically manages the \code{group_until} and
\code{pseudo_spatial} arguments.}

\item{pseudo_spatial}{a character or numeric vector giving the \emph{names or
indices} (NOT the column names to join on) of aggsequence that should have
'psuedo-spatial' aggregation. This is when we go from one spatial data
level to another, but do the join and aggregation with a non-spatial
\code{\link[dplyr:mutate-joins]{dplyr::left_join()}}. It is developed for the EWR situation, where the
incoming data is indexed to gauges, planning units, and sdl units, but has
gauge point geometry, and spatial joining to planning units or sdl is not
appropriate, because single gauges affect multiple units. So it would join
to the \code{planning_units} or \code{sdl_units} by column names instead of
spatially, and then aggregate according to those units. \emph{For EWR use,
typically the best option is \code{pseudo_spatial = c('planning_units', 'sdl_units')}, with the grouping to sdls only pseudo because some planning
units spill over sdl boundaries.}}

\item{returnList}{default \code{TRUE}, whether to return the output to the current
session}

\item{savepath}{default \code{NULL}, a path to save the output to. Note that this
names the output rds file directly '\code{type}_aggregated.rds', so the path
should include only the directory structure.  If \code{NULL}, does not save. If
\code{savepath = NULL} and \code{returnList = FALSE}, the function errors to avoid
wasting resources.}

\item{extrameta}{list, extra information to include in saved metadata
documentation for the run. Default NULL.}

\item{rparallel}{logical, default FALSE. If TRUE, parallelises over the
scenarios in hydro_dir using \code{furrr}. To use, install \code{furrr} and set a
\code{\link[future:plan]{future::plan()}} (likely \code{multisession} or \code{multicore})}

\item{par_recursive}{logical, default TRUE. If parallel, do we use the
innermost level of directory containing EWR outputs (TRUE) or the next
level in from datpath (FALSE)}

\item{savepar}{'combine' (default) or 'each'. If parallel over scenarios,
should this combine the output (default) or save each scenario's
aggregation separately ('each')}

\item{add_max, }{as in \code{\link[=read_and_geo]{read_and_geo()}} and \code{\link[=prep_ewr_output]{prep_ewr_output()}}, default TRUE.
Add a 'MAX' scenario that passes all EWRs, usable as a reference}

\item{...}{passed to \code{\link[=read_and_geo]{read_and_geo()}} and \code{\link[=prep_ewr_output]{prep_ewr_output()}}, primarily
\code{gaugefilter}, \code{scenariofilter}.}
}
\value{
either a tibble or sf of aggregated values at the final level (if
\code{saveintermediate = FALSE}) or a list of tibbles or sfs with aggregated
values at each step (\code{saveintermediate = TRUE})
}
\description{
Allows passing only paths to data rather than objects (though objects work as
well for consistency). Wrapper over \code{\link[=read_and_geo]{read_and_geo()}}, \code{\link[=make_edges]{make_edges()}} and
\code{\link[=multi_aggregate]{multi_aggregate()}}. Particularly useful if we want to pass parameters as
strings from a config before anything is read in, and parallelisation (set a
\code{\link[future:plan]{future::plan()}}).
}
