% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_and_geo.R
\name{read_and_geo}
\alias{read_and_geo}
\title{Prepare ewr results from the EWR tool for use in aggregation}
\usage{
read_and_geo(datapath, type, geopath, whichcrs = 4283, ...)
}
\arguments{
\item{datapath}{path to module results (Often EWR outputs with HydroBOT
modifications to format)}

\item{type}{character, a grep for the files to choose. The special case 'everything' gets all files}

\item{geopath}{sf object with geographic locations matching a column in the data, or path to a csv with gauge locations in lat/long (assumes BOM
currently) or a shapefile}

\item{whichcrs}{desired crs for the output}

\item{...}{passes \code{gaugefilter}, \code{scenariofilter} to \code{\link[=get_module_output]{get_module_output()}}.
Particularly useful for only processing some of the data at a time (e.g. in
parallel)}
}
\value{
an \code{sf} dataframe with the desired outputs including scenario
names and gauge or polygon locations as a POINT, POLYGON, or MULTIPOLYGON geometry. Primary purpose is to
prepare data to send to \code{\link[=multi_aggregate]{multi_aggregate()}}
}
\description{
This reads in the EWR results given a path and \code{type} (summary or annual),
and pairs them to their geographic information
}
