% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_toolkit_params.R
\name{run_toolkit_params}
\alias{run_toolkit_params}
\title{Run the toolkit through the aggregation step with a file of yaml parameters
or passing in yaml as characters.}
\usage{
run_toolkit_params(
  yamlpath = NULL,
  passed_args = NULL,
  list_args = NULL,
  defaults = system.file("yml/default_params.yml", package = "HydroBOT")
)
}
\arguments{
\item{yamlpath}{character, path to yaml file of parameters. Only needs to
contain values for parameters for which we wish to modify \code{defaults}. If we
want to be able to auto-run off the saved param metadata, the names of the
spatial agg levels need to match the objects, e.g. basin: basin}

\item{passed_args}{character, in yaml format, e.g. ""scenario_dir:
'project_dir/hydrographs'". Likely possible to pass more than one, but
complicated to get the yaml right. Suggest in that case modifying the
params.yml}

\item{list_args}{list of arguments. Typically would come in from
parameterised quarto notebook, which pre-parses yaml into a list}

\item{defaults}{character, path to default set of yaml parameters, so the
\code{yamlpath} file and \code{passed_args} only need to include those that are
modified}
}
\value{
runs the toolkit, returns NULL invisibly, or the aggregated output if
the params have \code{aggReturn: TRUE}
}
\description{
The plan is for this to make scripting easier, and be able to change single
simple parameters (e.g. directory) from command line or in other workflows
}
