% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ewr_output.R
\name{separate_ewr_codes}
\alias{separate_ewr_codes}
\title{Parser for EWR codes into the main code and the extra bits (called 'timing' for some reason)}
\usage{
separate_ewr_codes(df)
}
\arguments{
\item{df}{a dataframe with an ewr_code column with raw ewr names (e.g. EWR outputs, causal mappings)}
}
\value{
a dataframe with a clean ewr_code column and an ewr_code_timing column with the extra stuff
}
\description{
I want something that can be used against the whole ewr
}
