% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_toolkit_params.R
\name{type_cleanup}
\alias{type_cleanup}
\title{Sometimes single values come in as characters instead of lists}
\usage{
type_cleanup(arglist)
}
\arguments{
\item{arglist}{list of arguments}
}
\description{
Sometimes single values come in as characters instead of lists
}
