% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HydroBOT-package.R
\docType{package}
\name{HydroBOT-package}
\alias{HydroBOT}
\alias{HydroBOT-package}
\title{HydroBOT: Toolkit For Flow-Dependent Outcomes in MDB}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Galen Holt \email{g.holt@deakin.edu.au} (\href{https://orcid.org/0000-0002-7455-9275}{ORCID})

}
\keyword{internal}
